Added Ruby links to readme
