+++ Algorithms/Ruby/balanced-binary-tree.rb
