+++ Algorithms/Ruby/delete-kth-node.rb
