+++ Algorithms/Ruby/graph-depth-first-search.rb
