+++ Algorithms/Ruby/merge-sorted-arrays.rb
