+++ Algorithms/Ruby/permutation-palindrome.rb
