+++ Algorithms/Ruby/traverse-spiral-matrix.rb
