+++ Algorithms/Ruby/tree-breadth-first-search.rb
